+++ deployment/environment/crib/helpers.go
@@ -19,7 +19,7 @@ import (
 )
 
 const (
-	solFundingLamports = 10000000
+	solFundingLamports = 100000
 	evmFundingEth      = 100
 )
 
@@ -56,6 +56,7 @@ func distributeTransmitterFunds(lggr logger.Logger, nodeInfo []devenv.Node, env 
 
 	// Handle Solana funding
 	if len(env.SolChains) > 0 {
+		lggr.Info("Funding solana transmitters")
 		for sel, chain := range env.SolChains {
 			sel, chain := sel, chain
 			g.Go(func() error {
@@ -67,6 +68,7 @@ func distributeTransmitterFunds(lggr logger.Logger, nodeInfo []devenv.Node, env 
 						return err
 					}
 					base58Addr := n.AccountAddr[chainID]
+					lggr.Debugf("Found %v solana transmitter address", base58Addr)
 
 					pk, err := solana.PublicKeyFromBase58(base58Addr)
 					if err != nil {
@@ -126,47 +128,4 @@ func SendFundsToAccounts(ctx context.Context, lggr logger.Logger, chain deployme
 		nonce++
 	}
 	return nil
-}
-
-func distributeTransmitterFundsSolana(lggr logger.Logger, nodeInfo []devenv.Node, env deployment.Environment) error {
-	g := new(errgroup.Group)
-
-	const solFundingLamports = 100000
-
-	for sel, chain := range env.SolChains {
-		sel, chain := sel, chain
-
-		g.Go(func() error {
-			var solanaAddrs []solana.PublicKey
-			for _, n := range nodeInfo {
-				chainID, err := chainsel.GetChainIDFromSelector(sel)
-				if err != nil {
-					lggr.Errorw("could not get chain id from selector", "selector", sel, "err", err)
-					return err
-				}
-				base58Addr := n.AccountAddr[chainID]
-				lggr.Infof("Solana acc %v", n.AccountAddr)
-
-				for chainID, addr := range n.AccountAddr {
-					lggr.Infof("Chain ID: %s, Account Address: %s\n", chainID, addr)
-				}
-
-				pk, err := solana.PublicKeyFromBase58(base58Addr)
-				if err != nil {
-					lggr.Errorw("error converting base58 to solana PublicKey", "err", err, "address", base58Addr)
-					return err
-				}
-				solanaAddrs = append(solanaAddrs, pk)
-			}
-
-			err := memory.FundSolanaAccounts(env.GetContext(), solanaAddrs, solFundingLamports, chain.Client)
-			if err != nil {
-				lggr.Errorw("error funding solana accounts", "err", err, "selector", sel)
-				return err
-			}
-			return nil
-		})
-	}
-
-	return g.Wait()
 }
